+++ src/sentry/android/android_sdk.cpp
@@ -278,7 +278,7 @@ void AndroidSDK::_add_default_attachments() {
 	Object *android_plugin = _get_android_plugin();
 	ERR_FAIL_NULL(android_plugin);
 
-	for (const Ref<SentryAttachment> &att : SENTRY_OPTIONS()->get_file_attachments()) {
+	for (const Ref<SentryAttachment> &att : SENTRY_OPTIONS()->get_default_attachments()) {
 		android_plugin->call(ANDROID_SN(addFileAttachment),
 				att->get_globalized_path(),
 				att->get_effective_filename(),