+++ src/sentry/javascript/javascript_sdk.cpp
@@ -248,7 +248,7 @@ void JavaScriptSDK::clear_attachments() {
 	ERR_FAIL_COND(!js_bridge());
 
 	// Reset file attachments from options (Vector uses copy-on-write).
-	file_attachments = SENTRY_OPTIONS()->get_file_attachments();
+	file_attachments = SENTRY_OPTIONS()->get_default_attachments();
 	js_bridge()->call("clearAttachments");
 }
 
@@ -273,7 +273,7 @@ void JavaScriptSDK::metrics_add_distribution(const String &p_name, double p_valu
 void JavaScriptSDK::init() {
 	ERR_FAIL_COND(!js_bridge());
 
-	file_attachments = SENTRY_OPTIONS()->get_file_attachments();
+	file_attachments = SENTRY_OPTIONS()->get_default_attachments();
 
 	JSObjectPtr before_send_callback = JSObject::create_callback(before_send_wasm_callback);